Cockroach 22.2 backport of #1960.

----

**sstable: include more blocks' stats in BlockBytes**

Previously, BlockBytes and BlockBytesInCache only included data blocks
and sometimes the top-level index block. Now it always includes the
top-level index block, the first-level index block and the filter block.

This is motivated by a latter commit that needs the inclusion of index
blocks for a test case that illustrates the effectiveness of an
optimization.

**db: add ExternalIter_NonOverlapping_SeekNextScan benchmark**

Add a new benchmark that illustrates the pathological behavior of an external
iterator that must repeatedly seek into exhausted sstables.

**db: enable TrySeekUsingNext after Next in external iters**

Enable the TrySeekUsingNext optimization in another case, specifically
for external iterators. Previously this optimization only applied if the
previous positioning operation was also a seek.

This commit also enables it during forward iteration of external
iterators if the seek key is strictly greater than the current iterator
position's user key. This allows the optimization to trigger during
workloads that perform SeekGEs with monotonically increasing keys,
interspersed with Nexts. This is the case for CockroachDB scans.

This optimization is not enabled for non-external iterators because it
is unclear whether there is a performance benefit in these case. This
optimization is particularly beneficial with external iterators, because
it avoids the pathological behavior of re-seeking exhausted sstable
iterators.

Informs cockroachdb/cockroach#88329.

```
name                                                                                  old time/op    new time/op    delta
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=1/forward-only=false-16         45.8µs ± 1%    46.4µs ± 3%     ~     (p=0.222 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=1/forward-only=true-16          46.0µs ± 1%    35.4µs ± 2%  -23.02%  (p=0.008 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=10/forward-only=false-16         532µs ± 1%     545µs ± 2%   +2.40%  (p=0.016 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=10/forward-only=true-16          533µs ± 1%     300µs ± 0%  -43.67%  (p=0.008 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=100/forward-only=false-16       6.27ms ± 4%    6.09ms ± 1%     ~     (p=0.151 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=100/files=100/forward-only=true-16        5.93ms ± 1%    4.66ms ± 1%  -21.53%  (p=0.008 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=1/forward-only=false-16       2.06ms ± 1%    2.15ms ± 1%   +4.12%  (p=0.008 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=1/forward-only=true-16        2.10ms ± 2%    0.64ms ± 0%  -69.68%  (p=0.008 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=10/forward-only=false-16      48.2ms ± 1%    48.9ms ± 2%     ~     (p=0.095 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=10/forward-only=true-16       48.5ms ± 3%     2.1ms ± 1%  -95.64%  (p=0.008 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=100/forward-only=false-16      461ms ± 1%     470ms ± 3%     ~     (p=0.095 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=10000/files=100/forward-only=true-16       461ms ± 2%      15ms ± 1%  -96.67%  (p=0.008 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=1/forward-only=false-16       256ms ± 1%     257ms ± 2%     ~     (p=0.421 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=1/forward-only=true-16        255ms ± 2%       7ms ± 1%  -97.38%  (p=0.008 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=10/forward-only=false-16      438ms ± 2%     446ms ± 2%     ~     (p=0.095 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=10/forward-only=true-16       444ms ± 3%      19ms ± 1%  -95.81%  (p=0.008 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=100/forward-only=false-16     5.26s ± 1%     5.37s ± 1%   +2.16%  (p=0.008 n=5+5)
ExternalIter_NonOverlapping_SeekNextScan/keys=100000/files=100/forward-only=true-16      5.26s ± 1%     0.11s ± 1%  -97.92%  (p=0.008 n=5+5)
```